✨ Add ProgressViewController instead of views
