+++ ACKImagePicker/ViewControllers/ImagePickerViewController.swift
@@ -175,6 +175,15 @@ class ImagePickerViewController: BaseViewController {
         
         PHPhotoLibrary.shared().register(self)
     }
+    
+    // MARK: - LoadingView
+    
+    private func showProgress() -> ProgressViewController {
+        let controller = ProgressViewController()
+        controller.modalPresentationStyle = .overCurrentContext
+        present(controller, animated: false)
+        return controller
+    }
 }
 
 extension ImagePickerViewController: UITableViewDataSource {
@@ -334,8 +343,7 @@ extension ImagePickerViewController: ACKImagePickerDelegate {
         // Disable interaction with current view
         currentViewController?.view.isUserInteractionEnabled = false
         
-        let (progressView, overlayView) = setupUI(in: currentViewController)
-        currentViewController?.startLoadingAnimation()
+        let progressView = showProgress()
         
         var images: [UIImage] = []
         let manager = PHImageManager.default()
@@ -345,12 +353,9 @@ extension ImagePickerViewController: ACKImagePickerDelegate {
         options.isNetworkAccessAllowed = true
         options.progressHandler = { progress, _, _, _ in
             DispatchQueue.main.async {
-                guard let progressView = progressView else { return }
                 let imageFraction: Float = 1 / Float(photos.count)
                 let newProgress = Float(progress) * imageFraction + Float(images.count) / Float(photos.count)
-                print(newProgress)
-                guard progressView.progress < newProgress else { return }
-                progressView.setProgress(newProgress, animated: true)
+                progressView.progress = CGFloat(newProgress)
             }
         }
         
@@ -370,42 +375,40 @@ extension ImagePickerViewController: ACKImagePickerDelegate {
         
         // call `onImagesPicked` block on main thread when whole group is finished
         group.notify(queue: .main) { [weak self] in
-            progressView?.removeFromSuperview()
-            overlayView?.removeFromSuperview()
             currentViewController?.view.isUserInteractionEnabled = true
-            self?.stopLoadingAnimation()
+            progressView.dismiss(animated: false)
             self?.onImagesPicked?(images)
         }
     }
     
-    private func setupUI(in currentViewController: BaseViewController?) -> (progressView: UIProgressView?, overlayView: UIView?) {
-        guard let currentViewController = currentViewController else { return (progressView: nil, overlayView: nil) }
-        let overlayView = UIView()
-        overlayView.backgroundColor = UIColor.black.withAlphaComponent(0.5)
-        overlayView.isUserInteractionEnabled = false
-        currentViewController.view.addSubview(overlayView)
-        overlayView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            overlayView.leadingAnchor.constraint(equalTo: currentViewController.view.leadingAnchor),
-            overlayView.trailingAnchor.constraint(equalTo: currentViewController.view.trailingAnchor),
-            overlayView.bottomAnchor.constraint(equalTo: currentViewController.view.bottomAnchor),
-            overlayView.topAnchor.constraint(equalTo: currentViewController.view.topAnchor),
-        ])
-        
-        let progressView = UIProgressView()
-        currentViewController.view.addSubview(progressView)
-        progressView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            progressView.heightAnchor.constraint(equalToConstant: 5),
-            progressView.leadingAnchor.constraint(equalTo: currentViewController.view.leadingAnchor),
-            progressView.trailingAnchor.constraint(equalTo: currentViewController.view.trailingAnchor),
-        ])
-        if #available(iOS 11.0, *) {
-            progressView.topAnchor.constraint(equalTo: currentViewController.view.safeAreaLayoutGuide.topAnchor).isActive = true
-        } else {
-            progressView.topAnchor.constraint(equalTo: currentViewController.topLayoutGuide.bottomAnchor).isActive = true
-        }
-        
-        return (progressView: progressView, overlayView: overlayView)
-    }
+//    private func setupUI(in currentViewController: BaseViewController?) -> (progressView: UIProgressView?, overlayView: UIView?) {
+//        guard let currentViewController = currentViewController else { return (progressView: nil, overlayView: nil) }
+//        let overlayView = UIView()
+//        overlayView.backgroundColor = UIColor.black.withAlphaComponent(0.5)
+//        overlayView.isUserInteractionEnabled = false
+//        currentViewController.view.addSubview(overlayView)
+//        overlayView.translatesAutoresizingMaskIntoConstraints = false
+//        NSLayoutConstraint.activate([
+//            overlayView.leadingAnchor.constraint(equalTo: currentViewController.view.leadingAnchor),
+//            overlayView.trailingAnchor.constraint(equalTo: currentViewController.view.trailingAnchor),
+//            overlayView.bottomAnchor.constraint(equalTo: currentViewController.view.bottomAnchor),
+//            overlayView.topAnchor.constraint(equalTo: currentViewController.view.topAnchor),
+//        ])
+//        
+//        let progressView = UIProgressView()
+//        currentViewController.view.addSubview(progressView)
+//        progressView.translatesAutoresizingMaskIntoConstraints = false
+//        NSLayoutConstraint.activate([
+//            progressView.heightAnchor.constraint(equalToConstant: 5),
+//            progressView.leadingAnchor.constraint(equalTo: currentViewController.view.leadingAnchor),
+//            progressView.trailingAnchor.constraint(equalTo: currentViewController.view.trailingAnchor),
+//        ])
+//        if #available(iOS 11.0, *) {
+//            progressView.topAnchor.constraint(equalTo: currentViewController.view.safeAreaLayoutGuide.topAnchor).isActive = true
+//        } else {
+//            progressView.topAnchor.constraint(equalTo: currentViewController.topLayoutGuide.bottomAnchor).isActive = true
+//        }
+//        
+//        return (progressView: progressView, overlayView: overlayView)
+//    }
 }
